requirements.txt

# google-generativeai>=0.3.0
# streamlit>=1.22.0
# python-dotenv>=0.19.0
# gTTS>=2.3.1
# SpeechRecognition>=3.8.1
# pydub>=0.25.1
# sounddevice>=0.4.6
# numpy>=1.19.5
# Pillow>=9.0.0

streamlit>=1.32.0
python-dotenv>=1.0.1
pandas>=2.2.2
gtts>=2.5.1
speechrecognition>=3.10.0
openai>=1.33.0
pillow>=10.2.0
numpy>=1.26.4
